refactor: update import paths and remove obsolete local dataset store
