+++ src/syft_rds/client/local_stores/dataset/managers/__init__.py
@@ -1,7 +1,7 @@
-from .files import DatasetFilesManager
-from .path import DatasetPathManager
-from .schema import DatasetSchemaManager
-from .url import DatasetUrlManager
+from syft_rds.client.local_stores.dataset.managers.files import DatasetFilesManager
+from syft_rds.client.local_stores.dataset.managers.path import DatasetPathManager
+from syft_rds.client.local_stores.dataset.managers.schema import DatasetSchemaManager
+from syft_rds.client.local_stores.dataset.managers.url import DatasetUrlManager
 
 __all__ = [
     "DatasetFilesManager",
