+++ src/syft_rds/server/routers/job_router.py
@@ -5,7 +5,7 @@ from syft_core import SyftBoxURL
 from syft_event import SyftEvents
 from syft_event.types import Request
 import yaml
-from syft_rds.models.models import (
+from syft_rds.models import (
     Dataset,
     GetAllRequest,
     GetOneRequest,
@@ -16,7 +16,7 @@ from syft_rds.models.models import (
     JobStatus,
 )
 from syft_rds.server.router import RPCRouter
-from syft_rds.server.user_file_service import UserFileService
+from syft_rds.server.services.user_file_service import UserFileService
 from syft_rds.store import YAMLStore
 from syft_rds.utils.name_generator import generate_name
 from syft_rds.utils.zip_utils import zip_to_bytes
